+++ packages/expo-constants/src/Constants.types.ts
@@ -3,13 +3,11 @@ import type {
   EASConfig as ManifestsEASConfig,
   ExpoGoConfig as ManifestsExpoGoConfig,
   NewManifest,
-  LegacyManifest,
   BareManifest,
   ManifestAsset as ManifestAssetForReExport,
   ManifestExtra as ManifestExtraForReExport,
   ClientScopingConfig as ClientScopingConfigForReExport,
   ExpoGoPackagerOpts as ExpoGoPackagerOptsForReExport,
-  ExpoClientConfig as ExpoClientConfigForReExport,
   // @ts-ignore -- optional interface, will gracefully degrade to `any` not installed
 } from 'expo-manifests';
 
@@ -105,8 +103,6 @@ export type EASConfig = ManifestsEASConfig;
 export type ClientScopingConfig = ClientScopingConfigForReExport;
 export type ExpoGoConfig = ManifestsExpoGoConfig;
 export type ExpoGoPackagerOpts = ExpoGoPackagerOptsForReExport;
-export type ExpoClientConfig = ExpoClientConfigForReExport;
-export type AppManifest = LegacyManifest;
 
 // @needsAudit @docsMissing
 export interface PlatformManifest {
@@ -189,7 +185,7 @@ export interface NativeConstants {
    * @deprecated Use `Constants.expoConfig` instead, which behaves more consistently across EAS Build
    * and EAS Update.
    */
-  manifest: LegacyManifest | BareManifest | null;
+  manifest: BareManifest | null;
   /**
    * Manifest for Expo apps using modern Expo Updates from a remote source, such as apps that
    * use EAS Update. Returns `null` in bare workflow and when `manifest` is non-null.
@@ -253,7 +249,7 @@ export interface Constants extends NativeConstants {
    * In certain cases accessing manifest via this property
    * suppresses important warning about missing manifest.
    */
-  __unsafeNoWarnManifest?: LegacyManifest | BareManifest;
+  __unsafeNoWarnManifest?: BareManifest;
   /**
    * @hidden
    * @warning do not use this property. Use `manifest2` by default.